+++ dom/xslt/base/txExpandedNameMap.cpp
@@ -1,7 +1,7 @@
-
-
-
-
+/* -*- Mode: C++; tab-width: 4; indent-tabs-mode: nil; c-basic-offset: 4 -*- */
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
 #include "txExpandedNameMap.h"
 #include "txCore.h"
@@ -16,17 +16,17 @@ class txMapItemComparator
     }
 };
 
-
-
-
-
-
-
-
+/**
+ * Adds an item, if an item with this key already exists an error is
+ * returned
+ * @param  aKey   key for item to add
+ * @param  aValue value of item to add
+ * @return errorcode
+ */
 nsresult txExpandedNameMap_base::addItem(const txExpandedName& aKey,
                                          void* aValue)
 {
-    uint32_t pos = mItems.IndexOf(aKey, 0, txMapItemComparator());
+    size_t pos = mItems.IndexOf(aKey, 0, txMapItemComparator());
     if (pos != mItems.NoIndex) {
         return NS_ERROR_XSLT_ALREADY_SET;
     }
@@ -41,23 +41,22 @@ nsresult txExpandedNameMap_base::addItem(const txExpandedName& aKey,
     return NS_OK;
 }
 
-
-
-
-
-
-
-
+/**
+ * Sets an item, if an item with this key already exists it is overwritten
+ * with the new value
+ * @param  aKey   key for item to set
+ * @param  aValue value of item to set
+ * @return errorcode
+ */
 nsresult txExpandedNameMap_base::setItem(const txExpandedName& aKey,
                                          void* aValue,
                                          void** aOldValue)
 {
     *aOldValue = nullptr;
-    uint32_t pos = mItems.IndexOf(aKey, 0, txMapItemComparator());
+    size_t pos = mItems.IndexOf(aKey, 0, txMapItemComparator());
     if (pos != mItems.NoIndex) {
         *aOldValue = mItems[pos].mValue;
         mItems[pos].mValue = aValue;
-        
         return NS_OK;
     }
 
@@ -71,14 +70,14 @@ nsresult txExpandedNameMap_base::setItem(const txExpandedName& aKey,
     return NS_OK;
 }
 
-
-
-
-
-
+/**
+ * Gets an item
+ * @param  aKey  key for item to get
+ * @return item with specified key, or null if no such item exists
+ */
 void* txExpandedNameMap_base::getItem(const txExpandedName& aKey) const
 {
-    uint32_t pos = mItems.IndexOf(aKey, 0, txMapItemComparator());
+    size_t pos = mItems.IndexOf(aKey, 0, txMapItemComparator());
     if (pos != mItems.NoIndex) {
         return mItems[pos].mValue;
     }
@@ -86,16 +85,16 @@ void* txExpandedNameMap_base::getItem(const txExpandedName& aKey) const
     return nullptr;
 }
 
-
-
-
-
-
-
+/**
+ * Removes an item, deleting it if the map owns the values
+ * @param  aKey  key for item to remove
+ * @return item with specified key, or null if it has been deleted
+ *         or no such item exists
+ */
 void* txExpandedNameMap_base::removeItem(const txExpandedName& aKey)
 {
     void* value = nullptr;
-    uint32_t pos = mItems.IndexOf(aKey, 0, txMapItemComparator());
+    size_t pos = mItems.IndexOf(aKey, 0, txMapItemComparator());
     if (pos != mItems.NoIndex) {
         value = mItems[pos].mValue;
         mItems.RemoveElementAt(pos);
